requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
llvmlite==0.24.0
matplotlib==2.2.2
noise==1.2.2
nose==1.3.7
numba==0.39.0
numpy==1.14.5
opensimplex==0.2
Pillow==5.1.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
scipy==1.1.0
six==1.11.0
